Changed the Long and Double types to Ints and Floats.

+++ node2vec_spark/src/main/scala/au/csiro/data61/randomwalk/efficient/RandomSample.scala
@@ -2,18 +2,18 @@ package au.csiro.data61.randomwalk.efficient
 
 import scala.util.Random
 
-case class RandomSample(nextDouble: () => Double = Random.nextDouble) extends Serializable {
+case class RandomSample(nextFloat: () => Float = Random.nextFloat) extends Serializable {
 
 
   /**
     *
     * @return
     */
-  final def sample(edges: Array[(Long, Double)]): (Long, Double) = {
+  final def sample(edges: Array[(Int, Float)]): (Int, Float) = {
 
     val sum = edges.foldLeft(0.0) { case (w1, (_, w2)) => w1 + w2 }
 
-    val p = nextDouble()
+    val p = nextFloat()
     var acc = 0.0
     for ((dstId, w) <- edges) {
       acc += w / sum
@@ -24,11 +24,11 @@ case class RandomSample(nextDouble: () => Double = Random.nextDouble) extends Se
     edges.head
   }
 
-  final def computeSecondOrderWeights(p: Double = 1.0,
-                                      q: Double = 1.0,
-                                      prevId: Long,
-                                      prevNeighbors: Array[(Long, Double)],
-                                      currNeighbors: Array[(Long, Double)]): Array[(Long, Double)
+  final def computeSecondOrderWeights(p: Float = 1.0f,
+                                      q: Float = 1.0f,
+                                      prevId: Int,
+                                      prevNeighbors: Array[(Int, Float)],
+                                      currNeighbors: Array[(Int, Float)]): Array[(Int, Float)
     ] = {
     currNeighbors.map { case (dstId, w) =>
       var unnormProb = w / q // Default is that there is no direct link between src and
@@ -52,11 +52,11 @@ case class RandomSample(nextDouble: () => Double = Random.nextDouble) extends Se
     * @param currNeighbors
     * @return
     */
-  final def secondOrderSample(p: Double = 1.0,
-                              q: Double = 1.0,
-                              prevId: Long,
-                              prevNeighbors: Array[(Long, Double)],
-                              currNeighbors: Array[(Long, Double)]): (Long, Double) = {
+  final def secondOrderSample(p: Float = 1.0f,
+                              q: Float = 1.0f,
+                              prevId: Int,
+                              prevNeighbors: Array[(Int, Float)],
+                              currNeighbors: Array[(Int, Float)]): (Int, Float) = {
     val newCurrentNeighbors = computeSecondOrderWeights(p, q, prevId, prevNeighbors, currNeighbors)
     sample(newCurrentNeighbors)
   }
